+++ Bricouli/src/main/java/com/demo/Bricouli/repositories/UserRepository.java
@@ -1,5 +1,7 @@
 package com.demo.Bricouli.repositories;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,6 +10,8 @@ import com.demo.Bricouli.models.User;
 
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Double> {
+public interface UserRepository extends JpaRepository<User, Long> {
+	
+	
 
 }
